requirements.txt
google-api-python-client==2.123.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
pytest==8.1.1


